backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
authlib>=1.3.0
httpx>=0.26.0
whoosh>=2.7.4
